requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.16.0
djangorestframework-stubs==3.15.3
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
typing_extensions==4.13.2
urllib3==2.4.0
